monitoring_requirements.txt
# Study Room Monitoring System - Python Dependencies (Production)

# Core ML/CV Libraries (Lightweight)
opencv-python==4.8.1.78
mediapipe==0.10.15
numpy==1.24.3

# MediaPipe runtime dependencies (required for container)
onnxruntime>=1.17.0
absl-py>=1.0.0
protobuf>=3.20.0
attrs>=21.1.0

# API & Database
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.0.3
python-dotenv==1.0.0

# Scheduling & Background Tasks
apscheduler==3.10.4

# Processing & Utils
pillow==10.1.0
scipy==1.11.4
scikit-learn==1.3.2

# Monitoring & Logging
python-json-logger==2.0.7
psutil==5.9.6

# Build tools (required for some packages)
setuptools>=65.0.0
wheel

# Note: YOLOv8 (ultralytics) and PyTorch removed for lighter deployment
# These can be added later if needed for enhanced detection
# Optional: torch==2.0.1, ultralytics==8.0.212, torchvision==0.15.2

# Optional: GPU acceleration (uncomment if using NVIDIA GPU)
# cupy-cuda12x==12.3.0
# tensorrt==8.6.1

# Optional: Edge acceleration (uncomment for Coral TPU)
# tflite-runtime==2.13.0

# Optional: Mobile/Lite models
# tflite-support==0.4.4
